user login and child operations

+++ src/gql/resolvers/users.js
@@ -1,64 +1,83 @@
-const db = require( "../../../models" );
+const bcrypt = require( 'bcrypt' )
+const jwt = require( 'jsonwebtoken' )
+const { JWT_SECRET, JWT_EXPIRY } = require( "../../../const" )
 
-me = async ( parent, args, context, info ) => {
+const db = require( "../../../models" )
+
+const modelToJSON = ( mongooseModel ) => {
+  // console.log( 'User:', mongooseModel )
+  return {
+    id: mongooseModel._id,
+    ...mongooseModel.toJSON(),
+    password: '',
+  }
+}
+
+const me = async ( parent, args, context, info ) => {
   if ( !context.isLoggedIn ) throw '401 - user not available'
 
   const user = await db.User.findOne( { email: context.authUser.email } )
-  // console.log( context )
-  return {
-    id: user._id,
-    ...user.toJSON(),
-  }
+  return modelToJSON( user )
 }
 
-users = async ( parent, args, context, info ) => {
+const users = async ( parent, args, context, info ) => {
   if ( !context.isLoggedIn ) throw '401 - user not available'
 
   const users = await db.User.find( {} )
-  // console.log( context )
-  return users.map( user => ( {
-    id: user._id,
-    ...user.toJSON(),
-  } ) )
+  return users.map( modelToJSON )
 }
 
-user = async ( parent, args, context, info ) => {
+const user = async ( parent, args, context, info ) => {
   if ( !context.isLoggedIn ) throw '401 - user not available'
 
   const user = await db.User.findById( args.id )
-  // console.log( context )
-  return {
-    id: user._id,
-    ...user.toJSON(),
-  }
+  return modelToJSON( user )
 }
 
-updateUser = async ( parent, args, context, info ) => {
+const updateUser = async ( parent, args, context, info ) => {
+  if ( !context.isLoggedIn ) throw '401 - child not available'
+
   try {
     const { firstName, lastName, email } = args.data
     const user = await db.User.findByIdAndUpdate( args.id, {
       firstName, lastName, email
     }, { new: true } )
-    return {
-      id: user._id,
-      ...user.toJSON(),
-    }
+    return modelToJSON( user )
   } catch ( e ) {
     console.log( e )
     throw 'Unable to update the User'
   }
 }
 
-deleteUser = async ( parent, args, context, info ) => {
+const deleteUser = async ( parent, args, context, info ) => {
+  if ( !context.isLoggedIn ) throw '401 - child not available'
+
   try {
-    const user = db.User.findByIdAndDelete( args.id )
-    return {
-      id: user._id,
-      ...user.toJSON(),
-    }
+    const user = await db.User.findByIdAndDelete( args.id )
+    return modelToJSON( user )
   } catch ( e ) {
     throw 'Unable to delete the User'
   }
+}
+
+const login = async ( parent, args, context, info ) => {
+  const { email, password } = args
+
+  const foundUser = await db.User.findOne( { email } )
+  if ( !foundUser ) throw Error( 'User not found' )
+
+  const matched = await bcrypt.compare( password, foundUser.get( 'password' ) )
+  if ( !matched ) throw Error( 'Password mismatch' )
+
+  const user = modelToJSON( foundUser )
+  // console.log( user )
+  const token = jwt.sign( user, JWT_SECRET, { expiresIn: JWT_EXPIRY } )
+
+  return {
+    user,
+    token,
+  }
+
 }
 
 module.exports = {
@@ -66,9 +85,11 @@ module.exports = {
     me,
     user,
     users,
+    login,
   },
   Mutation: {
     updateUser,
     deleteUser,
-  }
+  },
+  modelToJSON,
 }
